Chore: using postgresql instead sqlite 🔨🔨
