+++ .env
@@ -3,3 +3,5 @@ JWT_SECRET=secret
 POSTGRES_DB='legends-back'
 POSTGRES_USER='admin'
 POSTGRES_PASSWORD='00cc645d-03b7-4b92-9554-f85c71afcea9'
+POSTGRES_HOST='localhost'
+POSTGRES_PORT=5432
