+++ src/setVarsJest.js
@@ -3,11 +3,9 @@
 process.env.port = 4000
 process.env.JWT_SECRET = 'secret'
 
-const { EntityModel, LogModel, UserModel } = require('./models')
+const { EntityModel, UserModel } = require('./models')
 
 // Initialize the database
 EntityModel.sync({ force: true })
-
-LogModel.sync({ force: true })
 
 UserModel.sync({ force: true })
